Tweaks

+++ postgres-backup-s3/backup.sh
@@ -2,6 +2,7 @@
 
 set -eo pipefail
 
+# Check required environment variables
 if [ "${S3_ACCESS_KEY_ID}" = "**None**" ]; then
   echo "You need to set the S3_ACCESS_KEY_ID environment variable."
   exit 1
@@ -38,30 +39,34 @@ if [ "${POSTGRES_USER}" = "**None**" ]; then
 fi
 
 if [ "${POSTGRES_PASSWORD}" = "**None**" ]; then
-  echo "You need to set the POSTGRES_PASSWORD environment variable or link to a container named POSTGRES."
+  echo "You need to set the POSTGRES_PASSWORD environment variable."
   exit 1
 fi
 
+# Configure AWS CLI options
 if [ "${S3_ENDPOINT}" == "**None**" ]; then
   AWS_ARGS=""
 else
   AWS_ARGS="--endpoint-url ${S3_ENDPOINT}"
 fi
 
-# env vars needed for aws tools
+# Set AWS environment variables
 export AWS_ACCESS_KEY_ID=$S3_ACCESS_KEY_ID
 export AWS_SECRET_ACCESS_KEY=$S3_SECRET_ACCESS_KEY
 export AWS_DEFAULT_REGION=$S3_REGION
 
+# Set Postgres environment variables
 export PGPASSWORD=$POSTGRES_PASSWORD
 POSTGRES_HOST_OPTS="-h $POSTGRES_HOST -p $POSTGRES_PORT -U $POSTGRES_USER $POSTGRES_EXTRA_OPTS"
 
+# Handle S3 prefix
 if [ -z ${S3_PREFIX+x} ]; then
   S3_PREFIX="/"
 else
   S3_PREFIX="/${S3_PREFIX}/"
 fi
 
+# Backup all databases or specific ones
 if [ "${POSTGRES_BACKUP_ALL}" == "true" ]; then
   SRC_FILE=dump.sql.gz
   DEST_FILE=all_$(date +"%Y-%m-%dT%H:%M:%SZ").sql.gz
@@ -71,23 +76,34 @@ if [ "${POSTGRES_BACKUP_ALL}" == "true" ]; then
   fi
 
   echo "Creating dump of all databases from ${POSTGRES_HOST}..."
-  pg_dumpall -h $POSTGRES_HOST -p $POSTGRES_PORT -U $POSTGRES_USER | gzip > $SRC_FILE
+  if pg_dumpall -h $POSTGRES_HOST -p $POSTGRES_PORT -U $POSTGRES_USER | gzip > $SRC_FILE; then
+    echo "Database dump successful."
+  else
+    echo "Database dump failed." >&2
+    exit 2
+  fi
 
   if [ "${ENCRYPTION_PASSWORD}" != "**None**" ]; then
     echo "Encrypting ${SRC_FILE}"
-    openssl enc -aes-256-cbc -in $SRC_FILE -out ${SRC_FILE}.enc -k $ENCRYPTION_PASSWORD
-    if [ $? != 0 ]; then
-      >&2 echo "Error encrypting ${SRC_FILE}"
+    if openssl enc -aes-256-cbc -in $SRC_FILE -out ${SRC_FILE}.enc -k $ENCRYPTION_PASSWORD; then
+      echo "Encryption successful."
+      rm $SRC_FILE
+      SRC_FILE="${SRC_FILE}.enc"
+      DEST_FILE="${DEST_FILE}.enc"
+    else
+      echo "Error encrypting ${SRC_FILE}" >&2
+      exit 2
     fi
-    rm $SRC_FILE
-    SRC_FILE="${SRC_FILE}.enc"
-    DEST_FILE="${DEST_FILE}.enc"
   fi
 
   echo "Uploading dump to $S3_BUCKET"
-  cat $SRC_FILE | aws $AWS_ARGS s3 cp - "s3://${S3_BUCKET}${S3_PREFIX}${DEST_FILE}" || exit 2
+  if cat $SRC_FILE | aws $AWS_ARGS s3 cp - "s3://${S3_BUCKET}${S3_PREFIX}${DEST_FILE}"; then
+    echo "SQL backup uploaded successfully"
+  else
+    echo "Error uploading to S3" >&2
+    exit 2
+  fi
 
-  echo "SQL backup uploaded successfully"
   rm -rf $SRC_FILE
 else
   OIFS="$IFS"
@@ -104,23 +120,34 @@ else
     fi
     
     echo "Creating dump of ${DB} database from ${POSTGRES_HOST}..."
-    pg_dump $POSTGRES_HOST_OPTS $DB | gzip > $SRC_FILE
+    if pg_dump $POSTGRES_HOST_OPTS $DB | gzip > $SRC_FILE; then
+      echo "Database dump successful."
+    else
+      echo "Database dump failed." >&2
+      exit 2
+    fi
     
     if [ "${ENCRYPTION_PASSWORD}" != "**None**" ]; then
       echo "Encrypting ${SRC_FILE}"
-      openssl enc -aes-256-cbc -in $SRC_FILE -out ${SRC_FILE}.enc -k $ENCRYPTION_PASSWORD
-      if [ $? != 0 ]; then
-        >&2 echo "Error encrypting ${SRC_FILE}"
+      if openssl enc -aes-256-cbc -in $SRC_FILE -out ${SRC_FILE}.enc -k $ENCRYPTION_PASSWORD; then
+        echo "Encryption successful."
+        rm $SRC_FILE
+        SRC_FILE="${SRC_FILE}.enc"
+        DEST_FILE="${DEST_FILE}.enc"
+      else
+        echo "Error encrypting ${SRC_FILE}" >&2
+        exit 2
       fi
-      rm $SRC_FILE
-      SRC_FILE="${SRC_FILE}.enc"
-      DEST_FILE="${DEST_FILE}.enc"
     fi
 
     echo "Uploading dump to $S3_BUCKET"
-    cat $SRC_FILE | aws $AWS_ARGS s3 cp - "s3://${S3_BUCKET}${S3_PREFIX}${DEST_FILE}" || exit 2
+    if cat $SRC_FILE | aws $AWS_ARGS s3 cp - "s3://${S3_BUCKET}${S3_PREFIX}${DEST_FILE}"; then
+      echo "SQL backup uploaded successfully"
+    else
+      echo "Error uploading to S3" >&2
+      exit 2
+    fi
 
-    echo "SQL backup uploaded successfully"
     rm -rf $SRC_FILE
   done
 fi
